// src/app/core/interfaces/endereco.interface.ts
export interface EnderecoInterface {
    cep: string;
    tipo_logradouro: number;
    logradouro: string;
    numero: number;
    bairro: string;
    complemento: string;
    municipio: number;
    uf: number;
    email: string;
    ddd_celular: string;
    celular: string;
    ddd_telefone: string;
    telefone: string;
}